renamed RequestMapping value on Controller add first "/bootadmin/entities"
