+++ src/main/java/ru/geekbrains/coursework/webshop/app/external/pages/bootadmin/entities/BrandController.java
@@ -1,4 +1,4 @@
-package ru.geekbrains.coursework.webshop.app.external;
+package ru.geekbrains.coursework.webshop.app.external.pages.bootadmin.entities;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -6,6 +6,6 @@ import ru.geekbrains.coursework.webshop.app.domain.BrandService;
 import ru.geekbrains.coursework.webshop.app.domain.entities.Brand;
 
 @Controller
-@RequestMapping("/brand")
+@RequestMapping("/bootadmin/entities/brand")
 public class BrandController extends AController<Brand, BrandService> {
 }
