+++ src/main/java/ru/geekbrains/coursework/webshop/app/external/pages/bootadmin/entities/CategoryController.java
@@ -1,4 +1,4 @@
-package ru.geekbrains.coursework.webshop.app.external;
+package ru.geekbrains.coursework.webshop.app.external.pages.bootadmin.entities;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -6,6 +6,6 @@ import ru.geekbrains.coursework.webshop.app.domain.CategoryService;
 import ru.geekbrains.coursework.webshop.app.domain.entities.Category;
 
 @Controller
-@RequestMapping("/category")
+@RequestMapping("/bootadmin/entities/category")
 public class CategoryController extends AController<Category, CategoryService> {
 }
