+++ src/main/java/ru/geekbrains/coursework/webshop/app/external/pages/bootadmin/entities/ProductController.java
@@ -1,9 +1,10 @@
-package ru.geekbrains.coursework.webshop.app.external;
+package ru.geekbrains.coursework.webshop.app.external.pages.bootadmin.entities;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import ru.geekbrains.coursework.webshop.app.domain.BrandService;
 import ru.geekbrains.coursework.webshop.app.domain.CategoryService;
@@ -11,7 +12,7 @@ import ru.geekbrains.coursework.webshop.app.domain.ProductService;
 import ru.geekbrains.coursework.webshop.app.domain.entities.Product;
 
 @Controller
-@RequestMapping("product")
+@RequestMapping("/bootadmin/entities/product")
 public class ProductController extends AController<Product, ProductService> {
     private CategoryService categoryService;
     private BrandService brandService;
@@ -23,8 +24,8 @@ public class ProductController extends AController<Product, ProductService> {
     }
 
     @Override
-    @GetMapping({"/show/{id}"})
-    public String show(Model model, Long id) {
+    @GetMapping("/show/{id}")
+    public String show(Model model, @PathVariable("id") Long id) {
         model.addAttribute("allBrands", brandService.getAll());
         model.addAttribute("allCategories", categoryService.getAll());
         return super.show(model, id);
