+++ src/main/java/ru/geekbrains/coursework/webshop/app/external/pages/bootadmin/entities/RoleController.java
@@ -1,4 +1,4 @@
-package ru.geekbrains.coursework.webshop.app.external;
+package ru.geekbrains.coursework.webshop.app.external.pages.bootadmin.entities;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -6,6 +6,6 @@ import ru.geekbrains.coursework.webshop.app.domain.RoleService;
 import ru.geekbrains.coursework.webshop.app.domain.entities.Role;
 
 @Controller
-@RequestMapping("role")
+@RequestMapping("/bootadmin/entities/role")
 public class RoleController extends AController<Role, RoleService> {
 }
